Android: Refactor Encoding handling and favourites
Encoding uses now a sharedPreference.
This breaks the Ini support but works for Zip archives.

Favourites use now a StringSet to make the Api simpler.
