+++ builds/android/app/src/main/java/org/easyrpg/player/game_browser/Game.java
@@ -15,9 +15,7 @@ public class Game implements Comparable<Game> {
     private String savePath;
 	private boolean isFavorite;
     private final DocumentFile gameFolder;
-    private Uri iniFile;
     private Bitmap titleScreen;
-    private IniFileManager iniFileManager; // Always use initIniFileManager() before using iniFileManager
     /** Path to the game folder inside of the zip */
     private String zipInnerPath;
 
@@ -26,30 +24,14 @@ public class Game implements Comparable<Game> {
 	    this.title = gameFolder.getName();
         Uri folderURI = gameFolder.getUri();
 	    this.gameFolderPath = folderURI.toString();
-
-		// SavePath
         this.savePath = gameFolderPath;
-        // TODO : Adapt this code in order to have savefile with .zip games
-        // TODO : Save file should be stored in the EasyRPGFolder (there is no more default folder)
-        /*
-        if (GameBrowserHelper.canWrite(f)) {
-			this.savePath = gameFolderPath;
-		} else {
-			// Not writable, redirect to a different path
-			// Try preventing collisions by using the names of the two parent directories
-			String savename = f.getParentFile().getName() + "/" + f.getName();
-			savePath = SettingsManager.getEasyRPGFolder() + "/Save/" + savename;
-			new File(savePath).mkdirs();
-		}
-        */
 
-		this.isFavorite = isFavoriteFromSettings();
+        this.isFavorite = isFavoriteFromSettings();
 	}
 
-	public Game(DocumentFile gameFolder, Uri iniFileUri, Bitmap titleScreen) {
+	public Game(DocumentFile gameFolder, Bitmap titleScreen) {
 	    this(gameFolder);
-        this.iniFile = iniFileUri;
-	    this.titleScreen = titleScreen;
+        this.titleScreen = titleScreen;
     }
 
     /**
@@ -66,13 +48,13 @@ public class Game implements Comparable<Game> {
         this.isFavorite = false;
     }
 
-    private Game(DocumentFile gameFolder, String pathInZip, Uri iniFileUri, Bitmap titleScreen) {
-        this(gameFolder, iniFileUri, titleScreen);
+    private Game(DocumentFile gameFolder, String pathInZip, Bitmap titleScreen) {
+        this(gameFolder, titleScreen);
         zipInnerPath = pathInZip;
     }
 
-    public static Game fromZip(DocumentFile gameFolder, String pathInZip, String saveFolder, Uri iniFileUri, Bitmap titleScreen) {
-        Game game = new Game(gameFolder, pathInZip, iniFileUri, titleScreen);
+    public static Game fromZip(DocumentFile gameFolder, String pathInZip, String saveFolder, Bitmap titleScreen) {
+        Game game = new Game(gameFolder, pathInZip, titleScreen);
         // is only relative here, launchGame will put this in the "saves" directory
         game.setSavePath(saveFolder);
         return game;
@@ -101,9 +83,9 @@ public class Game implements Comparable<Game> {
 	public void setFavorite(boolean isFavorite) {
 		this.isFavorite = isFavorite;
 		if(isFavorite){
-			SettingsManager.addFavoriteGame(this.title);
+			SettingsManager.addFavoriteGame(this);
 		} else {
-			SettingsManager.removeAFavoriteGame(this.title);
+			SettingsManager.removeAFavoriteGame(this);
 		}
 	}
 
@@ -122,25 +104,12 @@ public class Game implements Comparable<Game> {
 		return this.title.compareTo(game.title);
 	}
 
-    /** We initiate the IniFileManager only if necessary to prevent from unnecessary syscall */
-    public void initIniFileManager(Context context) {
-        if (iniFileManager == null) {
-            this.iniFileManager = new IniFileManager(context, iniFile);
-        }
-    }
-
-	public IniFileManager.Encoding getEncoding(Context context) {
-        // We initiate the IniFileManager only if necessary to prevent from unnecessary syscalls */
-        initIniFileManager(context);
-
-        return this.iniFileManager.getEncoding();
+	public Encoding getEncoding() {
+        return SettingsManager.getGameEncoding(this);
 	}
 
-	public void setEncoding(Context context, IniFileManager.Encoding encoding) {
-        // We initiate the IniFileManager only if necessary to prevent from unnecessary syscalls */
-        initIniFileManager(context);
-
-        this.iniFileManager.setEncoding(context, encoding);
+	public void setEncoding(Encoding encoding) {
+        SettingsManager.setGameEncoding(this, encoding);
 	}
 
     @NonNull
